Fixed Issues
Fixed cloning issue by creating new class VariableArray for arrays of T that extend Variable<T[]> and clonning the inside value.

Separated methods of Bnf_Reader for structured genetic evolution in new class.

Tried to fix some parsing issues that appeared in sonarCloud

+++ src/main/java/jeco/core/algorithm/es/EvolutionaryStrategy.java
@@ -69,7 +69,7 @@ public class EvolutionaryStrategy<T extends Variable<?>> extends Algorithm<T> {
 		int nextPercentageReport = 10;
 		while (currentGeneration < maxGenerations) {
 			step();
-			int percentage = Math.round((currentGeneration * 100) / maxGenerations);
+			int percentage = Math.round((currentGeneration * 100) / (float)maxGenerations);
 			if (percentage == nextPercentageReport) {
 				logger.info(percentage + "% performed ...");
 				nextPercentageReport += 10;
